Update file structure of image functions
In this commit, the image.py file is removed and the functions are moved
to a new image class.
